--- Assets/02.Scripts/All Inventory/Item Inventory/ItemInventory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dd95ee2bdaaa5454bad7c00d09a7a626
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 200
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
