+++ Assets/02.Scripts/All Inventory/Skill Inventory/SkillData/SkillSystem.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SkillSystem
+{
+    private SkillData[] skilldatas;
+
+    public SkillSystem()
+    {
+        skilldatas = SaveSys.LoadAllData().SkillDB;
+
+        foreach(var a in skilldatas)
+        {
+            a.Init();
+        }
+    }
+
+
+
+}
