+++ Assets/02.Scripts/All Inventory/Skill Inventory/SkillData/SkillSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fffbd2e72bbd58349b6603440f6c390b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
