--- Assets/02.Scripts/Dialog/DialogData.cs
@@ -1,8 +1,0 @@
-﻿[System.Serializable]
-public class DialogData
-{
-    public int nQuestID;
-    public int nBranch;
-    public string sName;
-    public string sDialog;
-}
