--- Assets/02.Scripts/Dialog/DialogData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2aac38bb5cdf2a94493c866e2b335bb1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
