+++ Assets/02.Scripts/Objects/Base/ObjectData.cs
@@ -28,6 +28,7 @@ public class ObjectData
         Init();
     }
 
+
     public ObjectData(ObjectBase objBase)
     {
         nLevel  = objBase.GetLevel();
