+++ Assets/02.Scripts/Objects/Monster/BabyDragon.cs
@@ -19,10 +19,35 @@ public class BabyDragon : Lv01_09
 
     protected override void Die()
     {
-        playerCtr.AddInven(_itemDic);
+        _playerCtr.AddInven(_itemDic);
 
         base.Die();
     }
 
+    protected override void LoadData()
+    {
+        //몬스터는 100번대 부터 시작한다. 0번째 인덱스 부터 시작하려면 m_ID - 100 
+        //현재 몬스터에 해당하는 INDEX를 찾아간다.
+        _monsterData = SaveSys.LoadAllData().MonsterDB[m_nID - 100];
+
+        if (_monsterData == null)
+        {
+            Debug.LogError("MonData is NULL!!");
+            return;
+        }
+
+
+
+        m_nLevel = _monsterData.nLevel;
+        m_nMaxHP = _monsterData.nMaxHP;
+        m_nCurHP = _monsterData.nMaxHP;
+
+        m_nMaxMP = _monsterData.nMaxMP;
+        m_nCurMP = _monsterData.nMaxMP;
+
+
+        m_nCurSTR = _monsterData.nCurSTR;
+        m_nCurExp = _monsterData.nDropExp;
+    }
 
 }
