--- Assets/02.Scripts/Quest/QuestData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d4a98e54ffd9b2f4089a2d9d9caf3333
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
