+++ Assets/02.Scripts/Quest/QuestUI.cs
@@ -4,15 +4,22 @@ using UnityEngine;
 using UnityEngine.UI;
 using TMPro;
 using System.Text;
+using System;
 
 public class QuestUI : MonoBehaviour
 {
+
+    [SerializeField] private int n_maxManageableQuest = 3;
     [SerializeField] private GameObject _QuestPanel;
+
+    [SerializeField] private TMP_Text[] _questHeaders;
+    [SerializeField] private TMP_Text[] _disCriptions;
+    private int[] _questSlotIDArray;
+
+    private int _playerLevel;
 
     private TMP_Text _questHeader;
     private TMP_Text _disCription;
-
-    private int _playerLevel;
 
     private QuestSystem _questSystem;
     private List<QuestData> _currentQuest;
@@ -22,10 +29,27 @@ public class QuestUI : MonoBehaviour
 
     private void Awake()
     {
+        _questHeaders = new TMP_Text[n_maxManageableQuest];
+        _disCriptions = new TMP_Text[n_maxManageableQuest];
+        _questSlotIDArray = new int[n_maxManageableQuest];
+
         _resourcesData = new ResourcesData();
 
         _questHeader = _resourcesData.GetQuestHeader();
         _disCription = _resourcesData.GetQuestConText();
+
+        //제목, 내용 생성과 할당.
+        for(int i = 0; i < n_maxManageableQuest; i++)
+        {
+            //퀘스트 패널의 자식으로 생성한다.
+            _questHeaders[i] = Instantiate<TMP_Text>(_questHeader, _QuestPanel.transform);
+            _disCriptions[i] = Instantiate<TMP_Text>(_disCription, _QuestPanel.transform);
+
+            //생성 후 비활성화
+            _questHeaders[i].gameObject.SetActive(false);
+            _disCriptions[i].gameObject.SetActive(false);
+        }
+
     }
 
 
@@ -37,64 +61,37 @@ public class QuestUI : MonoBehaviour
 
     public void UpdateQuestUI(List<QuestData> currentQuest)
     {
-        ClearPanel();
-
         int size = currentQuest.Count;
         int preID = 0;
         StringBuilder sb = new StringBuilder();
         for (int i = 0; i < size; i++)
         {
-            if(preID != currentQuest[i].nQuestID)
+            //초기 퀘스트 슬롯 index 세팅
+            if(_questSlotIDArray[i] == 0)
             {
-                string questName = currentQuest[i].sQuestName;
-
-                //중복 헤더 Name은 생략한다.
-
-                if (currentQuest[i].eQuestType == 1)
-                    sb.Append($"<color=orange>[메인]");
-                else
-                    sb.Append($"[일반]");
-
-
-                string name = questName;
-                sb.Append(name);
-
-                var headText = Instantiate<TMP_Text>(_questHeader, _QuestPanel.transform);
-                headText.text = sb.ToString();
-                sb.Clear();
-
-                preID = currentQuest[i].nID;
+                _questSlotIDArray[i] = currentQuest[i].nQuestID;
+                _questHeaders[i].gameObject.SetActive(true);
+                _disCriptions[i].gameObject.SetActive(true);
             }
 
-            // 현재 퀘스트 설명
-            string discription = currentQuest[i].sDiscription;
+            int index = Array.IndexOf(_questSlotIDArray, currentQuest[i].nQuestID);
 
-            //완료된 퀘스트라면 회색 밑줄 처리
-            if (currentQuest[i].bIsComplete)
-            {
-                sb.Append($"<s><color=grey>{discription}</s>");
-            }
+
+            Debug.Assert(index != -1, "Quest Slot index NULL");
+
+            if (currentQuest[i].eQuestType == 1)
+                sb.Append($"<color=orange>[메인]");
             else
-            {
-                sb.Append($"{discription}");
-            }
+                sb.Append($"[일반]");
 
-            var discriptionText = Instantiate<TMP_Text>(_disCription, _QuestPanel.transform);
-            discriptionText.text = sb.ToString();
+            sb.Append($"{currentQuest[i].sQuestName}");
+            _questHeaders[index].text = sb.ToString();
 
             sb.Clear();
+
+
+            sb.Append($"{currentQuest[i].sDiscription}");
+            _disCriptions[index].text = sb.ToString();
         }
     }
-
-    public void ClearPanel()
-    {
-        int count = _QuestPanel.transform.childCount;
-
-        for(int i = 0; i < count; i++)
-        {
-            Destroy(_QuestPanel.transform.GetChild(i).gameObject);
-        }
-
-    }
-
 }
