+++ Assets/GameRPGDB.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f236f5d0f4d063948a2dcfe9e039db92
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
